Change the type of fibheap key to uint32_t to speedup functions.
git-svn-id: https://shiva.ms.mff.cuni.cz/svn/zlomekfs/trunk@2138 b158dfce-25fb-0310-9a5b-f29f65ed45b5

+++ zfsd/fibheap.h
@@ -48,7 +48,7 @@
 #include "pthread.h"
 #include "alloc-pool.h"
 
-typedef uint64_t fibheapkey_t;
+typedef uint32_t fibheapkey_t;
 #define FIBHEAPKEY_MIN 0
 #define FIBHEAPKEY_MAX ((fibheapkey_t) -1)
